+++ tests/Model/YamlIndent/YamlIndentCheckerTest.php
@@ -13,7 +13,7 @@ class YamlIndentCheckerTest extends TestCase
 {
     public function testCheckUnfixedFileIsNotCorrect(): void
     {
-        $standardParametersData = $this->getStandardsParametersData();
+        $standardParametersData = $this->getStandardsParametersData(YamlStandardConfigDefinition::CONFIG_PARAMETERS_INDENTS_COMMENTS_WITHOUT_PARENT_VALUE_DEFAULT);
         $pathToFile = __DIR__ . '/resource/unfixed/yaml-getting-started.yml';
 
         $yamlIndentChecker = new YamlIndentChecker();
@@ -24,7 +24,7 @@ class YamlIndentCheckerTest extends TestCase
 
     public function testCheckUnfixedFilesIsNotCorrect(): void
     {
-        $standardParametersData = $this->getStandardsParametersData();
+        $standardParametersData = $this->getStandardsParametersData(YamlStandardConfigDefinition::CONFIG_PARAMETERS_INDENTS_COMMENTS_WITHOUT_PARENT_VALUE_DEFAULT);
         $pathToFiles = [
             __DIR__ . '/resource/unfixed/kustomization.yaml',
             __DIR__ . '/resource/unfixed/shopsys-service.yml',
@@ -45,8 +45,19 @@ class YamlIndentCheckerTest extends TestCase
 
     public function testCheckFixedFileIsCorrect(): void
     {
-        $standardParametersData = $this->getStandardsParametersData();
+        $standardParametersData = $this->getStandardsParametersData(YamlStandardConfigDefinition::CONFIG_PARAMETERS_INDENTS_COMMENTS_WITHOUT_PARENT_VALUE_DEFAULT);
         $pathToFile = __DIR__ . '/resource/fixed/yaml-getting-started.yml';
+
+        $yamlIndentChecker = new YamlIndentChecker();
+        $result = $yamlIndentChecker->check($pathToFile, $standardParametersData);
+
+        $this->assertSame(Result::RESULT_CODE_OK, $result->getResultCode());
+    }
+
+    public function testCheckFixedFileIsCorrectWithPreservedComments(): void
+    {
+        $standardParametersData = $this->getStandardsParametersData(YamlStandardConfigDefinition::CONFIG_PARAMETERS_INDENTS_COMMENTS_WITHOUT_PARENT_VALUE_PRESERVED);
+        $pathToFile = __DIR__ . '/resource/fixed/yaml-getting-started-with-preserved-comments.yml';
 
         $yamlIndentChecker = new YamlIndentChecker();
         $result = $yamlIndentChecker->check($pathToFile, $standardParametersData);
@@ -56,7 +67,7 @@ class YamlIndentCheckerTest extends TestCase
 
     public function testCheckFixedFilesIsCorrect(): void
     {
-        $standardParametersData = $this->getStandardsParametersData();
+        $standardParametersData = $this->getStandardsParametersData(YamlStandardConfigDefinition::CONFIG_PARAMETERS_INDENTS_COMMENTS_WITHOUT_PARENT_VALUE_DEFAULT);
         $pathToFiles = [
             __DIR__ . '/resource/fixed/kustomization.yaml',
             __DIR__ . '/resource/fixed/shopsys-service.yml',
@@ -76,10 +87,11 @@ class YamlIndentCheckerTest extends TestCase
     }
 
     /**
+     * @param string $indentsCommentsWithoutParent
      * @return \YamlStandards\Model\Config\StandardParametersData
      */
-    private function getStandardsParametersData(): StandardParametersData
+    private function getStandardsParametersData(string $indentsCommentsWithoutParent): StandardParametersData
     {
-        return new StandardParametersData(4, 4, 4, YamlStandardConfigDefinition::CONFIG_PARAMETERS_SERVICE_ALIASING_TYPE_VALUE_SHORT);
+        return new StandardParametersData(4, 4, 4, YamlStandardConfigDefinition::CONFIG_PARAMETERS_SERVICE_ALIASING_TYPE_VALUE_SHORT, $indentsCommentsWithoutParent);
     }
 }